+++ simplezakka/src/main/java/com/example/simplezakka/dto/order/CustomerInfo.java
@@ -12,11 +12,8 @@ import jakarta.validation.constraints.Size;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CustomerInfo {
-
     
-    
-    
-    private Integer customerId;
+    private Long customerId;
 
     @NotBlank(message = "氏名は必須です。")
     @Size(max = 100, message = "氏名は100文字以内で入力してください。")
